separate bids.js -> individual test files
