+++ validators/bids/index.js
@@ -1,3 +1,17 @@
-var bids = require('./bids')
+var start = require('./start')
+var reset = require('./reset')
+var quickTest = require('./quickTest')
+var quickTestError = require('./quickTestError')
+var fullTest = require('./fullTest')
+var mismatchTest = require('./sesMismatch')
 
-module.exports = bids
+module.exports = {
+  options: {},
+  issues: [],
+  start: start,
+  quickTestError: quickTestError,
+  quickTest: quickTest,
+  fullTest: fullTest,
+  subIDsesIDmismatchtest: mismatchTest,
+  reset: reset,
+}
