+++ validators/bids/reset.js
@@ -1,0 +1,10 @@
+/**
+ * Reset
+ *
+ * Resets the BIDS issue data back to original values.
+ */
+function reset(BIDS) {
+  BIDS.issues = []
+}
+
+module.exports = reset
